Move longitude_slicer to utils and fix float32 tolerance
Moves longitude_slicer from _source_bathy to utils where it belongs
as a general-purpose coordinate utility. Also loosens the uniform-spacing
assert (atol=1e-4, rtol=0) to handle float32 source data (e.g. GLORYS)
where accumulated rounding produces two slightly different diff values
for what is conceptually a uniform grid.

Closes #99
